+++ android/app/src/main/java/com/foss/foss/FossApplication.kt
@@ -4,6 +4,7 @@ import com.boogiwoogi.woogidi.application.DiApplication
 import com.boogiwoogi.woogidi.pure.Instance
 import com.foss.foss.data.repository.DefaultMatchRepository
 import com.foss.foss.data.repository.DefaultUserRepository
+import com.foss.foss.di.auto.RemoteDataSourceModule
 import com.foss.foss.repository.MatchRepository
 import com.foss.foss.repository.UserRepository
