+++ android/app/src/main/java/com/foss/foss/di/manual/ViewModelFactoryProvider.kt
@@ -1,4 +1,4 @@
-package com.foss.foss.di
+package com.foss.foss.di.manual
 
 import androidx.lifecycle.ViewModelProvider
 import androidx.lifecycle.viewmodel.initializer
